pyproject.toml
[project]
name = "congressapi-client"
version = "0.1.1"
description = "Typed Congress.gov API wrapper"
authors = [{ name="Sean Peters"}]
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }

dependencies = [
  "requests>=2.31",
  "tqdm>=4.67.1",
  "xmltodict>=0.14.2",
  "python-dotenv>=1.1.1",
  "notebook>=7.4.5",
  "pytest-timeout>=2.4.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.4.1",
  "requests-mock>=1.12.1",
]

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[dependency-groups]
dev = [
    "isort>=6.0.1",
    "pre-commit>=4.3.0",
]
